Extract image_featured + image_grid; refactor generate_image imports
Move fullscreen featured-game rendering to image_featured.py and
legacy linescore/grid code to image_grid.py. generate_image.py
re-exports both for backward compatibility.
